+++ NEWS.md
@@ -1,7 +1,7 @@
 # metadat 1.1-0 (2021-10-13)
 
 - added some more info to `dat.knapp2017`
-- added `dat.lehmann2018`, `dat.michael2013`, and `dat.mccurdy2020`
+- added `dat.lehmann2018`, `dat.michael2013`, `dat.mccurdy2020`, `dat.baker2009`, and `dat.bakdash2021`
 - changed concept term 'mean differences' to 'raw mean differences' (to better distinguish it from the concept 'standardized mean differences')
 
 # metadat 1.0-0 (2021-08-20)